requirements.txt
ipykernel
jupyter
matplotlib
tensorflow
tensorflow_hub
kagglehub
diffusers
huggingface
huggingface_hub
transformers
accelerate
torch
torchvision
Pillow
numpy
safetensors
python-dotenv
sentencepiece